# WEB/public/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Jersey Bola Retro API",
        "description": "API untuk aplikasi e-commerce Jersey Bola Retro. Menyediakan endpoint untuk manajemen produk, kategori, autentikasi, dan upload gambar.",
        "version": "1.0.0",
        "contact": {
            "name": "Jersey Bola Retro",
            "email": "admin@jerseyretro.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Development Server"
        }
    ],
...
    "paths": {
        "/api/auth/signup": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registrasi User Baru",
                "description": "Membuat akun user baru. Email admin@gmail.com atau admin@example.com akan otomatis mendapat role admin.",
                "operationId": "signUp",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignUpRequest"
                            },
                            "example": {
                                "name": "John Doe",
                                "email": "john@example.com",
                                "password": "password123"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SignUpResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Email sudah terdaftar atau data tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Daftar Semua Produk",
                "description": "Mengambil daftar semua produk yang tersedia. Mendukung filter berdasarkan kategori dan pencarian.",
                "operationId": "getProducts",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Kata kunci pencarian (nama, klub, deskripsi)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Manchester"
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter berdasarkan slug kategori",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "home"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar produk berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductListResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/products": {
            "get": {
                "tags": [
                    "Admin Products"
                ],
                "summary": "Daftar Produk (Admin)",
                "description": "Mengambil daftar semua produk dengan detail lengkap untuk admin",
                "operationId": "getAdminProducts",
                "responses": {
                    "200": {
                        "description": "Daftar produk berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductListResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin Products"
                ],
                "summary": "Tambah Produk Baru",
                "description": "Membuat produk baru dengan gambar dan ukuran",
                "operationId": "createProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProductRequest"
                            },
                            "example": {
                                "name": "Manchester United 1999 Home",
                                "slug": "manchester-united-1999-home",
                                "club": "Manchester United",
                                "season": "1998-1999",
                                "description": "Jersey klasik Manchester United treble 1999",
                                "price": 250000,
                                "categoryId": "uuid-category-id",
                                "imageUrl": "https://example.com/image.jpg",
                                "sizes": {
                                    "S": 5,
                                    "M": 10,
                                    "L": 10,
                                    "XL": 5,
                                    "XXL": 3
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produk berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Data tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/upload": {
            "post": {
                "tags": [
                    "Upload"
                ],
                "summary": "Upload Gambar",
                "description": "Upload file gambar ke Supabase Storage. Mendukung JPEG, PNG, WebP, dan GIF dengan ukuran maksimal 5MB.",
                "operationId": "uploadImage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary",
                                        "description": "File gambar (JPEG, PNG, WebP, GIF, max 5MB)"
                                    }
                                },
                                "required": [
                                    "file"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "File berhasil diupload",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UploadResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "File tidak valid (format/ukuran)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Upload error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/create-admin": {
            "post": {
                "tags": [
                    "Utilities"
                ],
                "summary": "Buat Admin Baru",
                "description": "Membuat user admin baru dengan Supabase Admin API",
                "operationId": "createAdmin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAdminRequest"
                            },
                            "example": {
                                "email": "admin@example.com",
                                "password": "admin123",
                                "name": "Admin"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Admin berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateAdminResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Email/password tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "description": "Pesan error"
                    }
                }
            },
            "SignUpRequest": {
                "type": "object",
                "required": [
                    "email",
                    "password",
                    "name"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "description": "Email user"
                    },
                    "password": {
                        "type": "string",
                        "minLength": 6,
                        "description": "Password (minimal 6 karakter)"
                    },
                    "name": {
                        "type": "string",
                        "description": "Nama lengkap user"
                    }
                }
            },
            "SignUpResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "User created successfully"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "name": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "user",
                            "admin"
                        ]
                    }
                }
            },
            "Product": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "club": {
                        "type": "string"
                    },
                    "season": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "price": {
                        "type": "number"
                    },
                    "stock": {
                        "type": "integer"
                    },
                    "isAvailable": {
                        "type": "boolean"
                    },
                    "categoryId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductImage"
                        }
                    },
                    "sizes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductSize"
                        }
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "ProductImage": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "url": {
                        "type": "string",
                        "format": "uri"
                    },
                    "alt": {
                        "type": "string"
                    },
                    "isPrimary": {
                        "type": "boolean"
                    },
                    "order": {
                        "type": "integer"
                    }
                }
            },
            "ProductSize": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "size": {
                        "type": "string",
                        "enum": [
                            "S",
                            "M",
                            "L",
                            "XL",
                            "XXL"
                        ]
                    },
                    "stock": {
                        "type": "integer"
                    }
                }
            },
            "Category": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    }
                }
            },
            "ProductListResponse": {
                "type": "object",
                "properties": {
                    "products": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Product"
                        }
                    }
                }
            },
            "ProductResponse": {
                "type": "object",
                "properties": {
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    }
                }
            },
...
            "CreateProductRequest": {
                "type": "object",
                "required": [
                    "name",
                    "slug",
                    "club",
                    "season",
                    "price",
                    "categoryId",
                    "imageUrl",
                    "sizes"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Nama produk"
                    },
                    "slug": {
                        "type": "string",
                        "description": "URL slug (unik)"
                    },
                    "club": {
                        "type": "string",
                        "description": "Nama klub"
                    },
                    "season": {
                        "type": "string",
                        "description": "Musim jersey (contoh: 1998-1999)"
                    },
                    "description": {
                        "type": "string",
                        "description": "Deskripsi produk"
                    },
                    "price": {
                        "type": "number",
                        "description": "Harga dalam Rupiah"
                    },
                    "categoryId": {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID kategori"
                    },
                    "imageUrl": {
                        "type": "string",
                        "format": "uri",
                        "description": "URL gambar produk"
                    },
                    "sizes": {
                        "type": "object",
                        "description": "Stok per ukuran",
                        "properties": {
                            "S": {
                                "type": "integer"
                            },
                            "M": {
                                "type": "integer"
                            },
                            "L": {
                                "type": "integer"
                            },
                            "XL": {
                                "type": "integer"
                            },
                            "XXL": {
                                "type": "integer"
                            }
                        }
                    }
                }
            },
...
            "UploadResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "url": {
                        "type": "string",
                        "format": "uri",
                        "description": "Public URL gambar"
                    },
                    "path": {
                        "type": "string",
                        "description": "Path file di storage"
                    }
                }
            },
            "CreateAdminRequest": {
                "type": "object",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    }
                }
            },
            "CreateAdminResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "user": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string",
                                "format": "uuid"
                            },
                            "email": {
                                "type": "string",
                                "format": "email"
                            },
                            "role": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    }
}